refactor(admin): split admin route into focused modules to reduce complexity
Extract Discord name refresh logic into adminRefresh.ts and user mutation
business logic into adminUserMutations.ts, leaving admin.ts as a thin HTTP
routing layer. Also converts handleChangeTwitchChannel to the parameter-object
pattern to address the function-parameters smell.
